+++ src/Jobs/ImportStages.php
@@ -13,6 +13,7 @@ use Matchish\ScoutElasticSearch\Jobs\Stages\PullFromSource;
 use Matchish\ScoutElasticSearch\Jobs\Stages\RefreshIndex;
 use Matchish\ScoutElasticSearch\Jobs\Stages\StageInterface;
 use Matchish\ScoutElasticSearch\Jobs\Stages\SwitchToNewAndRemoveOldIndex;
+use Matchish\ScoutElasticSearch\Jobs\Stages\WaitForImportRanges;
 use Matchish\ScoutElasticSearch\Searchable\ImportSource;
 
 /**
@@ -24,9 +25,10 @@ class ImportStages extends Collection
      * @param  ImportSource  $source
      * @param  bool  $parallel
      * @param  bool  $catchUp
+     * @param  bool  $watch  follow the range jobs and report their progress
      * @return self
      */
-    public static function fromSource(ImportSource $source, bool $parallel = false, bool $catchUp = false)
+    public static function fromSource(ImportSource $source, bool $parallel = false, bool $catchUp = false, bool $watch = false)
     {
         $index = Index::fromSource($source);
 
@@ -35,17 +37,20 @@ class ImportStages extends Collection
             // range job of the batch has finished. Everything above
             // them runs immediately, in the process that starts the
             // import.
+            $dispatch = new DispatchImportRanges($source, $index, new FinishImport(array_values(array_filter([
+                $catchUp ? new CatchUp($source, $index) : null,
+                new RefreshIndex($index),
+                new SwitchToNewAndRemoveOldIndex($source, $index),
+            ]))));
+
             /** @var array<StageInterface> $stages */
-            $stages = [
+            $stages = array_values(array_filter([
                 new CancelPreviousImport($source),
                 new CleanUp($source),
                 new CreateWriteIndex($source, $index),
-                new DispatchImportRanges($source, $index, new FinishImport(array_values(array_filter([
-                    $catchUp ? new CatchUp($source, $index) : null,
-                    new RefreshIndex($index),
-                    new SwitchToNewAndRemoveOldIndex($source, $index),
-                ])))),
-            ];
+                $dispatch,
+                $watch ? new WaitForImportRanges($dispatch) : null,
+            ]));
         } else {
             /** @var array<StageInterface> $stages */
             $stages = [